+++ app/Http/Controllers/Questao_ObjetivaController.php
@@ -15,8 +15,10 @@ class Questao_ObjetivaController extends Controller
     public function index()
     {
         //
-        $materias=\App\Materia::all();
-        return view('questao_objetiva',compact('materias') );
+        $professor_id = auth()->user()->id;
+
+        $disciplinas= \App\Professor::find($professor_id)->disciplinas()->get();
+        return view('questao_objetiva.questao_objetiva',compact('disciplinas') );
     }
 
     /**
@@ -27,7 +29,7 @@ class Questao_ObjetivaController extends Controller
     public function create()
     {
         //
-        return view('questao_objetiva');
+        return view('questao_objetiva.questao_objetiva');
     }
 
     /**
@@ -39,9 +41,12 @@ class Questao_ObjetivaController extends Controller
     public function store(StoreQuestaoObjetiva $request)
     {
         //
+        $professor_id = auth()->user()->id;
+
         $questao_objetiva = new \App\Questao_Objetiva();
-        $questao_objetiva->materia_id = $request->get('materia_id');
+        $questao_objetiva->disciplina_id = $request->get('disciplina_id');
         $questao_objetiva->assunto_id = $request->get('assunto_id');
+        $questao_objetiva->professor_id = $professor_id;
         $questao_objetiva->dificuldade = $request->get('dificuldade');
         $questao_objetiva->enunciado = $request->get('enunciado');
         $questao_objetiva->alternativa_correta = $request->get('correta');
@@ -81,16 +86,24 @@ class Questao_ObjetivaController extends Controller
     public function edit($id)
     {
         //
+        $professor_id = auth()->user()->id;
+
         $questao = \App\Questao_Objetiva::find($id);
-        $materias = \App\Materia::all();
-        $assunto=\App\Assunto::find($questao->assunto_id);
 
-        $assuntos_original = \App\Materia::find($questao->materia_id)->assuntos()->get();
-        $alternativas = $questao->alternativa()->get();
+        if ($questao->professor_id == $professor_id) {
+            $disciplinas= \App\Professor::find($professor_id)->disciplinas()->get();
+            $assunto=\App\Assunto::find($questao->assunto_id);
 
-        //dd($alternativas);
+            $assuntos_original = \App\Disciplina::find($questao->disciplina_id)->assuntos()->get();
+            $alternativas = $questao->alternativa()->get();
 
-        return view('editar_questao_objetiva',compact('questao' , 'id', 'materias', 'assunto', 'assuntos_original', 'alternativas'));
+            //dd($alternativas);
+
+            return view('questao_objetiva.editar_questao_objetiva',compact('questao' , 'id', 'disciplinas', 'assunto', 'assuntos_original', 'alternativas'));
+        }else{
+
+            return redirect('list_questoes_objetivas');
+        }
     }
 
     /**
@@ -100,30 +113,39 @@ class Questao_ObjetivaController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Storage $request, $id)
     {
         //
+        $professor_id = auth()->user()->id;
+
         $questao_objetiva = \App\Questao_Objetiva::find($id);
-        $questao_objetiva->materia_id = $request->get('materia_id');
-        $questao_objetiva->assunto_id = $request->get('assunto_id');
-        $questao_objetiva->dificuldade = $request->get('dificuldade');
-        $questao_objetiva->enunciado = $request->get('enunciado');
-        $questao_objetiva->alternativa_correta = $request->get('correta');
-        $questao_objetiva->save();
 
-        $questao_objetiva_id= $questao_objetiva->id;
-        $alternativa_id = $request->get('alternativa_id');
+        if ($questao_objetiva->professor_id == $professor_id) {
 
-        $alternativa = \App\Alternativa::find($alternativa_id);
-        $alternativa->alternativa_a = $request->get('altA');
-        $alternativa->alternativa_b = $request->get('altB');
-        $alternativa->alternativa_c = $request->get('altC');
-        $alternativa->alternativa_d = $request->get('altD');
-        $alternativa->alternativa_e = $request->get('altE');
-        $alternativa->questao_objetiva_id = $questao_objetiva_id;
-        $alternativa->save();
+            $questao_objetiva->disciplina_id = $request->get('disciplina_id');
+            $questao_objetiva->assunto_id = $request->get('assunto_id');
+            $questao_objetiva->professor_id = $professor_id;
+            $questao_objetiva->dificuldade = $request->get('dificuldade');
+            $questao_objetiva->enunciado = $request->get('enunciado');
+            $questao_objetiva->alternativa_correta = $request->get('correta');
+            $questao_objetiva->save();
 
-        return redirect('list_questoes_objetivas')->with('success','Questão atualizada com sucesso!');
+            $questao_objetiva_id= $questao_objetiva->id;
+            $alternativa_id = $request->get('alternativa_id');
+
+            $alternativa = \App\Alternativa::find($alternativa_id);
+            $alternativa->alternativa_a = $request->get('altA');
+            $alternativa->alternativa_b = $request->get('altB');
+            $alternativa->alternativa_c = $request->get('altC');
+            $alternativa->alternativa_d = $request->get('altD');
+            $alternativa->alternativa_e = $request->get('altE');
+            $alternativa->questao_objetiva_id = $questao_objetiva_id;
+            $alternativa->save();
+
+            return redirect('list_questoes_objetivas')->with('success','Questão atualizada com sucesso!');
+        }else{
+            return redirect('list_questoes_objetivas');
+        }
     }
 
     /**
@@ -135,37 +157,29 @@ class Questao_ObjetivaController extends Controller
     public function destroy($id)
     {
         //
+        $professor_id = auth()->user()->id;
         $questao = \App\Questao_Objetiva::find($id);
-        $questao->delete();
-        return redirect('list_questoes_objetivas')->with('success','Questão deletada!');
-    }
 
-    public function getAssuntos($materia_id)
-    {
-        //
-            $assuntos = \App\Materia::find($materia_id)->assuntos;
-            //$assuntos->get(['id', 'descricao']);
-            //return Response::json($assuntos);
-            return response()->json($assuntos);
-
-            
+        if ($questao->professor_id == $professor_id) {
+            $questao->delete();
+            return redirect('list_questoes_objetivas')->with('success','Questão deletada!');
+        }else{
+            return redirect('list_questoes_objetivas');
+        }
     }
 
     public function listar_questoes()
     {
         //
-        $questoes = \App\Questao_Objetiva::all();
+        $professor_id = auth()->user()->id;
+
+        $questoes = \App\Professor::find($professor_id)->questoes_objetivas()->get();
 
         foreach ($questoes as $key => $questao) {
-            $disciplinas[$key] = \App\Materia::find($questao->materia_id);
+            $disciplinas[$key] = \App\Disciplina::find($questao->disciplina_id);
             $assuntos[$key] = \App\Assunto::find($questao->assunto_id);
         }
 
-        //dump($questoes);
-        //dump($disciplinas);
-        //dd($assuntos);
-        //$disciplinas = \App\Materia::all();
-
-        return view('listar_questoes_objetivas',compact('questoes','disciplinas','assuntos'));
+        return view('questao_objetiva.listar_questoes_objetivas',compact('questoes','disciplinas','assuntos'));
     }
 }
